backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
pymongo
azure-core
azure-ai-inference